+++ find.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 // find node
-node * find ( list * lst, int key ) {
+node * find ( list * lst, ll key ) {
 	node * err;
 	if ( ( err = ( node * ) malloc ( sizeof ( node ) ) ) == NULL ) {
 		printf("Can't return ERROR NODE") ;
@@ -18,9 +18,9 @@ node * find ( list * lst, int key ) {
 	}
 	return p;
 }
-node * index_node ( list * lst , int i ) {
+node * index_node ( list * lst , ll i ) {
 	node * n = lst -> front -> next ;
-	register int k ;
+	register ll k ;
 	for ( k = 0 ; k < i ; k ++ ) {
 		if ( n == lst -> rear -> prev ) {
 			return n;
